backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
email-validator